um-197: edited tests logic
