+++ data/create_fixtures.py
@@ -12,12 +12,11 @@ from werkzeug.security import generate_password_hash
 
 from app import create_app, db
 from models.accounts import Security, User
-from models.patterns import SubCategoryEnumDict, get_subcategory
 from models.products import Categories, Product
 from models.shops import Shop
+from validation.products import SubCategoryEnum
 
 cur_dir = Path(__file__).resolve().parent
-
 
 def create_fixtures():
     with open("data/users_fixture.json") as f:
