+++ models/shops.py
@@ -12,6 +12,7 @@ from utils.utils import serialize
 SHOPS_PHOTOS_PATH = os.path.join(Config.MEDIA_PATH, 'shops')
 SHOPS_BANNER_PHOTOS_PATH = os.path.join(Config.MEDIA_PATH, 'banner_shops')
 
+
 class Shop(db.Model):
     __tablename__ = "shops"
 
@@ -108,15 +109,13 @@ class Shop(db.Model):
         shop_info = serialize(shop)
         # TODO: use object.get(attr) instead object[attr]+++++
         if shop_info.get("banner_shop") is not None:
-
             banner_shop_path = url_for('static', filename=f'media/'
-                                        f'banner_shop/{shop_info["banner_shop"]}',
-                                        _external=True)
+                                                          f'banner_shop/{shop_info["banner_shop"]}',
+                                       _external=True)
             shop_info['banner_shop'] = banner_shop_path
         if shop_info.get("photo_shop") is not None:
-
             photo_shop_path = url_for('static', filename=f'media/'
-                                        f'shops/{shop_info["photo_shop"]}',
-                                        _external=True)
+                                                         f'shops/{shop_info["photo_shop"]}',
+                                      _external=True)
             shop_info['photo_shop'] = photo_shop_path
         return shop_info
