+++ routes/products.py
@@ -1,4 +1,3 @@
-
 import logging
 
 from flask import Blueprint, jsonify, request
@@ -15,6 +14,7 @@ products = Blueprint("products_route", __name__, url_prefix="/products")
 
 CORS(products, supports_credentials=True)
 
+
 @products.route("/product", methods=["POST"])
 @jwt_required()
 def create_product():
@@ -30,34 +30,32 @@ def create_product():
     try:
         response = Product.add_product(**serialize_data)
         return jsonify(response), 201
-    except ValueError as e:
+    except (ValueError, UserError, NotFoundError) as e:
         return jsonify({'error': str(e)}), 400
-    except UserError as e:
-        return jsonify({'error': str(e)}), 400
-    except NotFoundError as e:
-        return jsonify({'error': str(e) }), 404
     except Exception as e:
         logging.error(e)
         return jsonify({'error': 'internal server error'}), 500
+
 
 @products.route("/product_photo/<int:product_id>", methods=["POST"])
 @jwt_required()
 def add_product_photo(product_id):
     if 'image' not in request.files:
         return jsonify({'error': 'No file provided'}), 400
-    try:      
+    try:
         raw_value = request.form.get('main', '').lower()
         main_photo = raw_value == 'true'
         photo = request.files['image']
-        response = ProductPhoto.add_product_photo(product_id, photo, main_photo) 
+        response = ProductPhoto.add_product_photo(product_id, photo, main_photo)
         return jsonify(response), 200
     except UserError as e:
         return jsonify({'error': str(e)}), 400
     except NotFoundError as e:
-        return jsonify({'error': str(e) }), 404
+        return jsonify({'error': str(e)}), 404
     except Exception as e:
         logging.error(e)
         return jsonify({'error': 'internal server error'}), 500
+
 
 @products.route("/shop_products", methods=["GET"])
 @jwt_required()
@@ -68,10 +66,11 @@ def get_shop_products():
     except ValueError as e:
         return jsonify({'error': str(e)}), 400
     except NotFoundError as e:
-        return jsonify({'error': str(e) }), 404
+        return jsonify({'error': str(e)}), 404
     except Exception as e:
         logging.error(e)
         return jsonify({'error': 'internal server error'}), 500
+
 
 @products.route("/update/<int:product_id>", methods=["PUT"])
 @jwt_required()
@@ -91,21 +90,22 @@ def update_product(product_id):
     except UserError as e:
         return jsonify({'error': str(e)}), 400
     except NotFoundError as e:
-        return jsonify({'error': str(e) }), 404
+        return jsonify({'error': str(e)}), 404
     except Exception as e:
         logging.error(e)
         return jsonify({'error': 'internal server error'}), 500
 
+
 @products.route("/deactivate/<int:product_id>", methods=["DELETE"])
 @jwt_required()
-def deactivate_product(product_id):    
+def deactivate_product(product_id):
     try:
         response = Product.delete_product(product_id)
         return jsonify(response), 200
     except UserError as e:
         return jsonify({'error': str(e)}, 400)
     except NotFoundError as e:
-        return jsonify({'error': str(e) }), 404
+        return jsonify({'error': str(e)}), 404
     except Exception as e:
         logging.error(e)
         return jsonify({'error': 'internal server error'}), 500
