+++ tests/conftest.py
@@ -14,7 +14,7 @@ from config.config import TestConfig
 # можна скопіювати json файли - покласти в парку tests/data
 # в json файли додати не валідні данні(кейси)
 # створити окремі функції в поточному файлі для завантаження тих чи інших тестових сеймплів(з json файлів)
-from data.create_fixtures import create_fixture_t
+from data.create_fixtures import create_fixtures
 from dependencies import db
 from tests import status
 
@@ -23,9 +23,15 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 N = 0
 
 
+class TestValidData:
+    TEST_EMAIL = "test@mail.com"
+    TEST_FULL_NAME = "TestName Full"
+    TEST_PASSWORD = "123467898qweW"
+
+
 def get_payload():
-# TODO
-# забрати глобальну змінну. Використовувати з функції описані вище
+    # TODO
+    # забрати глобальну змінну. Використовувати з функції описані вище
     global N
     N += 1
     payload = {
@@ -104,6 +110,7 @@ def prepopulated_engine(app):
         db.create_all()
         yield db
         db.drop_all()
+
 
 @pytest.fixture(scope="function")
 def prepopulated_session(prepopulated_engine):
